pyproject.toml
[tool.poetry]
name = "bus-bunching"
version = "0.1.0"
description = ""
authors = ["Joe Gambino <joe@gam.bingo>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
requests = "^2.28.1"
cta-bus-tracker = {git = "https://github.com/ideo/cta-bus-tracker.git", rev = "main"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
